Store tax IDs as ints rather than strings
Gives a nice increase in speed as well as using quite a bit less memory.

NCBI tax IDs are numeric and there are currently on the order of 100,000 of them.  The docs say "more than 100 new species are added to the database daily, and that the rate is accelerating"; however, OCaml ints use 63 bits on 64 bit systems, so we won't have to worry about not being able to store them.