Fix bug in the field splitter
There was a bug in the field splitter code, where the final field was always parsed as the empty string, even if it contained data. This was hidden because the tests also had a bug in which every generated line ended in the field delimiter followed by the end delimiter.  Because of this, both the test oracle and the splitter implementation returned the correct empty string for the final field.  However, the splitter would return an empty string for all final fields, regardless of whether they were actually non-empty.

This commit fixes the splitter, fixes the quickcheck tests, and adds more tests of edgecases.